refactor(tabs): rename .scss files to .module.scss
