+++ packages/tabs/src/TabList.tsx
@@ -1,7 +1,7 @@
 import { TabList as ReachTabList } from "@reach/tabs"
 import React, { ReactNode } from "react"
 import classnames from "classnames"
-import styles from "./styles.scss"
+import styles from "./Tabs.module.scss"
 
 export interface TabListProps {
   /**
