+++ packages/tabs/src/Tabs.module.scss
